move helper functions into testUtils - RSC-1158
